LPD-19822 RankingIndexWriter is an internal interface, it should not be registered as osgi service
